+++ src/main/java/com/sondplay/tweaks/SondPlayTweaks.java
@@ -19,5 +19,5 @@ import cpw.mods.fml.common.Mod;
 public class SondPlayTweaks {
     public static final String MODID   = "sondplaytweaks";
     public static final String NAME    = "SondPlayTweaks";
-    public static final String VERSION = "0.3.0";
+    public static final String VERSION = "0.4.0";
 }